More changes, comment out inactive backends for now, move things around
